modify: header 공통에서 개별로 변경
