--- src/app/(isHeader)/chat/page.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const ChatList = () => {
-  return <div>chatList</div>;
-};
-
-export default ChatList;
