--- src/app/(isHeader)/create/page.tsx
@@ -1,5 +1,0 @@
-const CreateMGC = () => {
-  return <div>create MGC</div>;
-};
-
-export default CreateMGC;
