--- src/app/(isHeader)/layout.tsx
@@ -1,17 +1,0 @@
-import { ReactNode } from 'react';
-import Header from '@/app/_components/Header';
-
-const IsHeaderLayout = ({
-  children,
-}: Readonly<{
-  children: ReactNode;
-}>) => {
-  return (
-    <>
-      <Header />
-      {children}
-    </>
-  );
-};
-
-export default IsHeaderLayout;
