--- src/app/(isHeader)/mypage/page.tsx
@@ -1,5 +1,0 @@
-const MyPage = () => {
-  return <div>myPage</div>;
-};
-
-export default MyPage;
