+++ src/app/chat/page.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { ChevronLeftIcon } from 'lucide-react';
+
+const ChatList = () => {
+  return (
+    <div>
+      <ChevronLeftIcon />
+      chatList
+    </div>
+  );
+};
+
+export default ChatList;
