+++ src/app/create/page.tsx
@@ -1,0 +1,12 @@
+import { ChevronLeftIcon } from 'lucide-react';
+
+const CreateMGC = () => {
+  return (
+    <div>
+      <ChevronLeftIcon />
+      create MGC
+    </div>
+  );
+};
+
+export default CreateMGC;
