+++ src/app/mgc/[id]/page.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
+import { ChevronLeftIcon } from 'lucide-react';
 
 const MGCDetail = () => {
   return (
     <div>
+      <ChevronLeftIcon />
       MGCDetail
       <div>MGCDetail</div>
       <div>MGCDetail</div>
