+++ src/app/mypage/page.tsx
@@ -1,0 +1,12 @@
+import { ChevronLeftIcon } from 'lucide-react';
+
+const MyPage = () => {
+  return (
+    <div>
+      <ChevronLeftIcon />
+      myPage
+    </div>
+  );
+};
+
+export default MyPage;
